Design

+++ install-dev.sh
@@ -24,10 +24,13 @@ cd ..
 rm -rf facebook-sdk
 
 # Redis installation
-curl -O http://download.redis.io/redis-stable.tar.gz
-tar xvzf redis-stable.tar.gz
-rm redis-stable.tar.gz
-cd redis-stable
-make
-cd ..
-mv redis-stable redis
+if [ ! -d redis/src ]; then
+    curl -O http://download.redis.io/redis-stable.tar.gz
+    tar xvzf redis-stable.tar.gz
+    rm redis-stable.tar.gz
+    cd redis-stable
+    make
+    cd ..
+    mv redis-stable redis
+fi
+
